refactor(vm2): move WsdbIpcMerkleDB next to its sibling DB impls
The WSDB IPC-backed LowLevelMerkleDBInterface lived at top-level
barretenberg/wsdb_client/ but it is purely an AVM-simulation primitive:

- Implements bb::avm2::simulation::LowLevelMerkleDBInterface.
- Its only sibling impls (HintingRawDB, PureRawMerkleDB,
  MockLowLevelMerkleDB) all live under vm2/simulation/{lib,testing}/ in
  the bb::avm2::simulation namespace.
- It already linked vm2_sim; the top-level placement was the outlier.

Now under barretenberg/cpp/src/barretenberg/vm2/simulation/dbs/, in
namespace bb::avm2::simulation. No consumers reference it yet (the IPC
stack cutover wires it in later), so this is a zero-blast-radius rename.

Kept as its own STATIC library (`wsdb_ipc_merkle_db`) rather than folded
into vm2_sim so nodejs_module / wasm builds don't pull in the IPC stack
they never use. The vm2/CMakeLists.txt globs that build vm2_sim and vm2
both explicitly exclude this source path; the library target is gated on
NOT(FUZZING) AND NOT(WASM) like the underlying wsdb_ipc_client.
