+++ barretenberg/cpp/src/CMakeLists.txt
@@ -130,7 +130,6 @@ if(NOT FUZZING AND NOT WASM AND NOT BB_LITE)
     # CMake binary dir kept under build/ so artifacts land alongside other libs.
     add_subdirectory(${CMAKE_SOURCE_DIR}/../../ipc-runtime/cpp ${CMAKE_BINARY_DIR}/ipc-runtime)
     add_subdirectory(barretenberg/wsdb)
-    add_subdirectory(barretenberg/wsdb_client)
     add_subdirectory(barretenberg/nodejs_module)
 endif()
